requirements.dev.txt
pytest==7.0.0
pre-commit==2.19.0
pycln==1.3.2
isort==5.10.1
black==22.3.0
flake8==4.0.1
pydocstyle==6.1.1
python-dotenv==0.21.0